refactor: 💡 abstract "load control" into its own module
